ticket VS-1598 fix of calculation rtp->avg_ptime

+++ rtp.h
@@ -359,7 +359,7 @@ public:
 	bool find_by_dest;
 	bool ok_other_ip_side_by_sip;
 	u_int16_t sdp_ptime;
-	u_int16_t avg_ptime;
+	float avg_ptime;
 	u_int32_t avg_ptime_count;
 	RtpGraphSaver graph;
 	#if not EXPERIMENTAL_SUPPRESS_AST_CHANNELS
